backend/requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
scikit-learn==1.4.1.post1
pandas==2.2.1
numpy==1.26.4
joblib==1.3.2
python-dotenv==1.0.1
requests==2.31.0
groq==0.4.2
